requirements.txt
python-pptx==0.6.23
tk==0.1.0